+++ include/rz_config.hpp
@@ -19,9 +19,9 @@ static const std::string PROJECT_DESCRIPTION = "Convert a JPG or PNG into all co
 
 static const std::string PROJECT_EXECUTABLE = "img2icon";
 
-static const std::string PROJECT_VERSION = "1.0.0";
+static const std::string PROJECT_VERSION = "1.1.0";
 static const std::int32_t PROJECT_VERSION_MAJOR { 1 };
-static const std::int32_t PROJECT_VERSION_MINOR { 0 };
+static const std::int32_t PROJECT_VERSION_MINOR { 1 };
 static const std::int32_t PROJECT_VERSION_PATCH { 0 };
 
 static const std::string PROJECT_HOMEPAGE_URL = "https://github.com/Zheng-Bote/img2icon";